Exception handling

+++ src/main/java/com/lucasdev/dscommerce/services/ProductService.java
@@ -3,54 +3,72 @@ package com.lucasdev.dscommerce.services;
 import com.lucasdev.dscommerce.dto.ProductDTO;
 import com.lucasdev.dscommerce.entities.Product;
 import com.lucasdev.dscommerce.repositories.ProductRepository;
+import com.lucasdev.dscommerce.services.exceptions.DatabaseExeception;
+import com.lucasdev.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
-
     @Autowired
     private ProductRepository repository;
 
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id) {
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
+    public ProductDTO findById(Long id){
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
+
         ProductDTO dto = new ProductDTO(product);
         return dto;
     }
 
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable pageable) {
+    public Page<ProductDTO> findAll(Pageable pageable){
         Page<Product> result = repository.findAll(pageable);
-        return result.map(x -> new ProductDTO(x));
+
+        return  result.map(product -> new ProductDTO(product));
     }
 
     @Transactional
-    public ProductDTO insert(ProductDTO dto) {
+    public ProductDTO insert(ProductDTO dto){
         Product entity = new Product();
         copyDtoToEntity(dto, entity);
+
         entity = repository.save(entity);
         return new ProductDTO(entity);
     }
 
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+    public ProductDTO update(Long id, ProductDTO dto){
+        try{
+            Product entity = repository.getReferenceById(id);
+
+            copyDtoToEntity(dto, entity);
+
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-       repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseExeception("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -59,5 +77,4 @@ public class ProductService {
         entity.setPrice(dto.getPrice());
         entity.setImgUrl(dto.getImgUrl());
     }
-
 }
